Schema: add some unique constraints (#74)
* Schema: add some unique constraints

The validation on `google_places_id` didn't allow for when it might be
`nil`, so tidied that up, too. It appears `unique` indexes allow for
arbitrary `null` values, so that can stay unchanged.

* Specs: introduce test fixtures

This adds test fixtures for currencies and users. This is a
controversial topic in the Rails testing space. Fixtures create some
"mystery guests" that are in all tests, which could conceivably cause
some confusion at times. In practice, I haven't seen this to be the
case. One of the major benefits of fixtures is in terms of test
performance. Rails inserts them in bulk at the beginning of the test
run, and transactional fixtures cleans them up after each individual
example, so you don't have the performance problems associated with
factories, which do database queries to insert on the fly.

That being said, I'd advocate for a combination of both factories and
fixtures. Fixtures are good for records that are going to be widely
used. We are going to want a `User` record in about a billion places and
won't care too much about the details, so why not already have one lined
up in the database? In other cases where we need a more customized
record, using the factory makes more sense.

https://semaphoreci.com/blog/2014/01/14/rails-testing-antipatterns-fixtures-and-factories.html
https://brandonhilkert.com/blog/7-reasons-why-im-sticking-with-minitest-and-fixtures-in-rails/
https://evilmartians.com/chronicles/factories-or-fixtures