Refactors: cache carousel on listing
Rather than caching each individual image, this caches all of the images
under the listing. If we want to re-fetch the images we can simply
`touch` the listing to invalidate the cache key.

We might want to reconsider this caching at some point, since it only
works for this particular template. If we end up wanting to re-use the
images in another place, we can probably find a way to cache inside the
model instead.