Refactors: pull `ENV` access into helper (#104)
Rather than accessing `ENV["GOOGLE_MAPS_API_KEY"]` in different places
this pulls it up into a helper method in `application_controller.rb`.
This way fewer places need to know where the information is coming from.
It can make things easier to test potentially by stubbing the value, and
avoids a global mystery guest, instead using a method call.

I also switched to using `.fetch`, which will throw an error if the key
is not found. This can help reduce head scratching as to why the app
isn't doing what we think it should. Because of this, we'll also need a
`.env.test.local` file, as `.env.local` does not get loaded in test
mode.