Upgrades: introduce chromedriver-helper
This makes it so that devs don't need to manually install `chromedriver`
and put it in their path somewhere. We might also consider the
`webdrivers` gem at some point if `geckodriver` becomes necessary.
`chromedriver-helper` is just bundled by default in new Rails apps so
figured I'd stick with that for now.

https://github.com/flavorjones/chromedriver-helper
https://github.com/titusfortner/webdrivers

I needed to update the chrome version on CircleCI, as the latest
chromedriver doesn't work with whatever version they have installed.

https://gist.github.com/Arjeno/8564d9643f16d072a85b9c9b5a9f7de0