Upgrades: remove version constraints for gems
These don't need to be locked down in `Gemfile`, as they are locked in
`Gemfile.lock`. As a general philosophy, I don't like to lock things
down in `Gemfile` unless there is an issue of some sort preventing us
from upgrading a particular gem, in which case I'll often add an
explanatory comment or link to an issue on Github for reference.