Seeds: add mock data for dev mode (#196)
We have `db/seeds.rb`, but I treat this as a file that can be run in
production, whereas we also want dev-only data, like a fake user to poke
around with. For this, I added a `db/mock_data.rb` file, and enhanced
the `rake db:reset` command to populate the database from this file. We
can also use `rake db:populate` to just add the records and nothing
else. I `break` early to prevent accidentally running this in
production.