Geocoder: enable ip location for dev/test mode (#206)
This adds a method in `ApplicationController` to fake `REMOTE_ADDR` for
geocoding based on IP address. The default IP is in San Francisco. If we
want to override it, we can do so with `REMOTE_ADDR=1.1.1.1 rails s` or
by adding it in `.env.local`.

Side note, according to the docs, `request.location` is slow due to it
hitting a 3rd party API, so we'll probably want to find a way to
background it at some point.

https://github.com/alexreisner/geocoder#geocoding-is-slow