Specs: enable `render_views` for all specs (#223)
By default, controller specs *only* run the controller code.
`render_views` causes them to also render the corresponding view, which
gives us a more integration-ey test, making sure that the variables we
set in the controller work with the views. They also make it so that we
can test against the view instead of checking the instance variables set
by the controller.

In this case, it surfaced a bug inside `ListingsController`. Because the
categories variables weren't set it was throwing an error.