Views: switch to using currency_ids for params (#238)
Passes back `currency_ids` and nests them under the `listing` params. I
needed to disable the validation on `CurrenciesListing` because it
validates before the listing saves, so because the listing hasn't been
saved, yet, the id will not be present. It will, however, be present
when the `CurrenciesListing` record is auto-saved after the `Listing`.

This is angling towards using the `collection_check_boxes` form helper.

https://apidock.com/rails/v4.0.2/ActionView/Helpers/FormBuilder/collection_check_boxes