Views: only require images for new records (#240)
This form validation makes you upload an image when updating a listing,
even when the listing already has an image. I opted to make it required
only for new listings, though another alternative would be to make it
`@listing.images.empty?`. The reason I chose this is that a user might
want to update the record for another reason, so doesn't make sense for
someone to have to upload an image in that case.