Views: add `satoshi_image` helper (#243)
Since we use this in so many places, seems worth pulling out rather than
copying and pasting it over and over. I made a judgement call on
`width:`, as in some places the width was different in the url than
specified for the image. Not sure if this was intentional, but I opted
to just use the url width.